+++ integration_tests/src/tests/mod.rs
@@ -1,0 +1,21 @@
+mod authenticate_message_test;
+pub use authenticate_message_test::*;
+mod batch_onboarding;
+pub use batch_onboarding::*;
+// mod batch_onboarding_deals_test;
+// mod change_beneficiary_test;
+// mod commit_post_test;
+// mod datacap_tests;
+// mod evm_test;
+// mod extend_sectors_test;
+// mod init_test;
+// mod market_miner_withdrawal_test;
+// mod multisig_test;
+// mod power_scenario_tests;
+// mod publish_deals_test;
+// mod replica_update_test;
+// mod terminate_test;
+// mod verified_claim_test;
+// mod verifreg_remove_datacap_test;
+mod withdraw_balance_test;
+pub use withdraw_balance_test::*;
